+++ controllers/auth/user-logout-controller.js
